pyproject.toml
[project]
name = "sql-unit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "daff>=1.4.2",
    "fire>=0.7.0",
    "jinja2>=3.1.6",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "sqlparse>=0.5.3",
]
scripts = { sql-unit = "sql_unit.cli:main" }


[dependency-groups]
dev = ["duckdb>=1.3.1", "pytest>=8.4.1"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
